+++ Example/SectionsViewController.swift
@@ -27,16 +27,17 @@ class SectionsViewController: UIViewController, DTCollectionViewManageable, UICo
     
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        self.manager.startManagingWithDelegate(self)
-        self.manager.registerCellClass(SolidColorCollectionCell.self)
-        self.manager.registerHeaderClass(SimpleTextCollectionReusableView.self)
-        self.manager.registerFooterClass(SimpleTextCollectionReusableView.self)
-        (self.collectionView?.collectionViewLayout as? UICollectionViewFlowLayout)?.headerReferenceSize = CGSize(width: 320, height: 50)
-        (self.collectionView?.collectionViewLayout as? UICollectionViewFlowLayout)?.footerReferenceSize = CGSize(width: 320, height: 50)
+    
+        manager.startManagingWithDelegate(self)
         
-        self.addSection()
-        self.addSection()
+        manager.registerCellClass(SolidColorCollectionCell.self)
+        manager.registerHeaderClass(SimpleTextCollectionReusableView.self)
+        manager.registerFooterClass(SimpleTextCollectionReusableView.self)
+        (collectionView?.collectionViewLayout as? UICollectionViewFlowLayout)?.headerReferenceSize = CGSize(width: 320, height: 50)
+        (collectionView?.collectionViewLayout as? UICollectionViewFlowLayout)?.footerReferenceSize = CGSize(width: 320, height: 50)
+        
+        addSection()
+        addSection()
     }
     
     @IBAction func addSection()
@@ -52,13 +53,13 @@ class SectionsViewController: UIViewController, DTCollectionViewManageable, UICo
     }
 
     @IBAction func removeSection(_ sender: AnyObject) {
-        if self.manager.memoryStorage.sections.count > 0 {
-            self.manager.memoryStorage.deleteSections(IndexSet(integer: manager.memoryStorage.sections.count - 1))
+        if manager.memoryStorage.sections.count > 0 {
+            manager.memoryStorage.deleteSections(IndexSet(integer: manager.memoryStorage.sections.count - 1))
         }
     }
     @IBAction func moveSection(_ sender: AnyObject) {
-        if self.manager.memoryStorage.sections.count > 0 {
-            self.manager.memoryStorage.moveSection(self.manager.memoryStorage.sections.count - 1, toSection:0)
+        if manager.memoryStorage.sections.count > 0 {
+            manager.memoryStorage.moveSection(manager.memoryStorage.sections.count - 1, toSection:0)
         }
     }
 }