+++ Source/SectionModel+ConvenienceGetters.swift
@@ -33,20 +33,20 @@ public extension SectionModel
     /// UICollectionView header model for current section
     var collectionHeaderModel : Any? {
         get {
-           return self.supplementaryModelOfKind(UICollectionElementKindSectionHeader)
+            return self.supplementaryModelOfKind(UICollectionElementKindSectionHeader, atIndex: 0)
         }
         set {
-            self.setSupplementaryModel(newValue, forKind: UICollectionElementKindSectionHeader)
+            self.setSupplementaryModel(newValue, forKind: UICollectionElementKindSectionHeader, atIndex: 0)
         }
     }
     
     /// UICollectionView footer model for current section
     var collectionFooterModel : Any? {
         get {
-            return self.supplementaryModelOfKind(UICollectionElementKindSectionFooter)
+            return self.supplementaryModelOfKind(UICollectionElementKindSectionFooter, atIndex: 0)
         }
         set {
-            self.setSupplementaryModel(newValue, forKind: UICollectionElementKindSectionFooter)
+            self.setSupplementaryModel(newValue, forKind: UICollectionElementKindSectionFooter, atIndex: 0)
         }
     }
 }